SysfsUtils: move to platform/linux and rename to SysfsPath
