+++ xbmc/platform/linux/CMakeLists.txt
@@ -1,11 +1,13 @@
 set(SOURCES CPUInfoLinux.cpp
             MemUtils.cpp
             OptionalsReg.cpp
+            SysfsPath.cpp
             TimeUtils.cpp)
 
 set(HEADERS CPUInfoLinux.h
             OptionalsReg.h
             PlatformConstants.h
+            SysfsPath.h
             TimeUtils.h)
 
 if(ALSA_FOUND)
