+++ xbmc/platform/linux/SysfsPath.h
